Wed, Nov 23, 2022, 1:58 PM +00:00

+++ filter-less/helpers.c
@@ -3,7 +3,7 @@
 // Convert image to grayscale
 void grayscale(int height, int width, RGBTRIPLE image[height][width])
 {
-    int gscale_average = 0;
+    double gscale_average = 0;
     for (int i = 0; i < width; i++)
     {
         for (int j = 0; j < height; j++)
